feat: record pid of init process
This commit adds recording of the pid of the init process. This is necessary in order to stop a
container, because we need to know the pid of the init process to send a signal. This commit is a
little **dirty** since mount /proc happens before chroot, which could lead to some security issues
when $rootfs/proc is a symbolic link.