refactor: add event delivery between init process and socker-shim
In the previous design, socker-shim is waiting for the init process to complete the entire
initialization process. This made it very difficult to implement net namespace initialization. This
commit uses a named pipe to pass events between socker-shim process and init process, so that
socker-shim process can assist the initiation of the init process.