feat: setup veth pair between host and container
This commit implements the communication between host and container through a pair of veth devices.
The veth device on the host side is named as veth${container_id:0:7}, and is joined into a Linux
bridge named vbridge_default. The veth device on the container side is renamed to eth0. The
assignment of ipv4 addresses is not yet implemented. But it is already able to ping using ipv6 local
link addresses.